+++ database/migrations/2019_10_01_115619_create_warehouse_images_table.php
@@ -15,7 +15,7 @@ class CreateWarehouseImagesTable extends Migration
     {
         Schema::create('warehouse_images', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('warehouse_id')->unsigned();
+            $table->bigInteger('warehouse_id')->unsigned();
             $table->string('source');
             $table->timestamps();
         });
